Enable flexible activations in tabular model
The current implementation of TabularModel forces the use of ReLU in all
layers. This is not always needed. In order to remedy this, a change was
implemented that allows for specifying other activation functions for
each layer. The implementation preserves the current functionality,
which defaults to ReLU when no activation function is given.

There are no check ahead of time, that the resulting model makes sense
or can even run. It is up to the user to ensure that the inclusion of
other activation functions makes sense.